Get rid of the pod pkg (yay!)
